agent_runtime/steward_ai_zorba_bot/requirements.txt
python-telegram-bot==21.3
python-dotenv==1.0.1
openai==1.6.1
httpx==0.27.0
PyYAML==6.0.3
jsonschema==4.23.0
google-auth==2.40.3
google-api-python-client==2.176.0
